We are provided with historical sales data for 45 Walmart stores located in different regions.
We are provided with historical sales data for 45 Walmart stores located in different regions. Each store contains many departments, and the goal is to project the sales for each department in each store.To add to the challenge, selected holiday markdown events are included in the dataset.These markdowns are known to affect sales, but it is challenging to predict which departments are affected and the extent of the impact.